+++ panel/tests/pane/test_media.py
@@ -1,10 +1,22 @@
 import pathlib
 
+from base64 import b64encode
+from io import BytesIO
+
 import numpy as np
+import pytest
+
+try:
+    from scipy.io import wavfile
+except Exception:
+    wavfile = None
 
 from panel.pane import Audio, Video
 
 ASSETS = pathlib.Path(__file__).parent / 'assets'
+
+scipy_available = pytest.mark.skipif(wavfile is None, reason="requires scipy")
+
 
 
 def test_video_url(document, comm):
@@ -50,3 +62,27 @@ def test_numpy_audio(document, comm):
     model = audio.get_root(document, comm=comm)
 
     assert model.value == 'data:audio/wav;base64,UklGRsQAAABXQVZFZm10IBAAAAABAAEAQB8AAIA+AAACABAAZGF0YaAAAAAAAF4ErQjgDOgQuBRDGH0bXB7WIOMifCScJT8mYyYHJi0l1yMLIs0fJR0dGr4WFBMqDw4LzQZ1Ahf+vfl59VfxZu2z6UrmN+OD4DjeXNz42g7aotm22UjaWNvi3ODeTOEe5E3nzuqU7pXywvYO+2v/yAMZCFAMXhA1FMoXDxv7HYMgoCJJJHolLyZlJhwmVSURJFciKiCTHZoaSBeqE8oP' # noqa
+
+
+@scipy_available
+def test_audio_array(document, comm):
+    data = np.random.randint(-100,100, 100).astype('int16')
+    sample_rate = 10
+    buffer = BytesIO()
+    wavfile.write(buffer, sample_rate, data)
+    b64_encoded = b64encode(buffer.getvalue()).decode('utf-8')
+
+    audio = Audio(data, sample_rate=sample_rate)
+    model = audio.get_root(document, comm=comm)
+    model_value = model.value
+
+    assert model_value.split(',')[1] == b64_encoded
+    assert model.value.startswith('data:audio/wav;base64')
+
+
+def test_audio_url(document, comm):
+    audio_url = 'http://ccrma.stanford.edu/~jos/mp3/pno-cs.mp3'
+    audio = Audio(audio_url)
+    model = audio.get_root(document, comm=comm)
+
+    assert audio_url == model.value
